+++ assets/js/phoenix_live_view/live_socket.js
@@ -174,11 +174,13 @@ export default class LiveSocket {
 
   isDebugEnabled(){ return this.sessionStorage.getItem(PHX_LV_DEBUG) === "true" }
 
+  isDebugDisabled(){ return this.sessionStorage.getItem(PHX_LV_DEBUG) === "false" }
+
   enableDebug(){ this.sessionStorage.setItem(PHX_LV_DEBUG, "true") }
 
   enableProfiling(){ this.sessionStorage.setItem(PHX_LV_PROFILE, "true") }
 
-  disableDebug(){ this.sessionStorage.removeItem(PHX_LV_DEBUG) }
+  disableDebug(){ this.sessionStorage.setItem(PHX_LV_DEBUG, "false") }
 
   disableProfiling(){ this.sessionStorage.removeItem(PHX_LV_PROFILE) }
 
@@ -198,6 +200,8 @@ export default class LiveSocket {
   getSocket(){ return this.socket }
 
   connect(){
+    // enable debug by default if on localhost and not explicitly disabled
+    if(window.location.hostname === "localhost" && !this.isDebugDisabled()){ this.enableDebug() }
     let doConnect = () => {
       if(this.joinRootViews()){
         this.bindTopLevelEvents()